CORE-460: Update Swift Demo App with NetworkFee Selection
Placeholders for Sweep and Payment
